Refactor FactoryBase partials
