+++ src/Dock.Model/FactoryBase.Creation.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Wiesław Šoltés. All rights reserved.
+// Copyright (c) Wiesław Šoltés. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 using System.Collections.Generic;
 using Dock.Model.Controls;
@@ -7,7 +7,7 @@ using Dock.Model.Core;
 namespace Dock.Model;
 
 /// <summary>
-/// Factory base class.
+/// Factory base class - creation methods.
 /// </summary>
 public abstract partial class FactoryBase
 {
@@ -76,4 +76,9 @@ public abstract partial class FactoryBase
 
     /// <inheritdoc/>
     public abstract IRootDock? CreateLayout();
+    /// <inheritdoc/>
+    public abstract IDictionary<IDockable, object> ToolControls { get; }
+
+    /// <inheritdoc/>
+    public abstract IDictionary<IDockable, object> DocumentControls { get; }
 }
